#554: move online services to src/onlineservice
