+++ src/layout/fallback_layout.cc
@@ -38,14 +38,14 @@
 
 #ifdef ONLINE_SERVICES
 
-#include "online_service.h"
+#include "onlineservice/online_service.h"
 
 #ifdef SOPCAST
-    #include "sopcast_content_handler.h"
+    #include "onlineservice/sopcast_content_handler.h"
 #endif
 
 #ifdef ATRAILERS
-    #include "atrailers_content_handler.h"
+    #include "onlineservice/atrailers_content_handler.h"
 #endif
 
 #endif//ONLINE_SERVICES
